Changes Commit

+++ DemoCommerce.API/Entity/Coupon.cs
@@ -9,15 +9,9 @@ namespace DemoCommerce.API.Entity
         public int CouponId { get; set; }
         public double DiscountAmount { get; set; }
         public required string CouponCode { get; set; }
-        public int MinAmount { get; set; }
+        public double MinAmount { get; set; }
 
 
-        //public class CouponConfig : IEntityTypeConfiguration<Coupon>
-        //{
-        //    public void Configure(EntityTypeBuilder<Coupon> builder)
-        //    {
-        //        //Configurations goes here
-        //    }
-        //}
+        
     }
 }
